CORE-153: Use NIO for browse: reversed type change that breaks serialization

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/uploader/AbstractUploader.java
@@ -289,7 +289,8 @@ public abstract class AbstractUploader implements Uploader {
 
     public void measureBandwidth() {
         long written = totalAmountRead + getAmountWritten();
-        session.measureBandwidth(written);
+        // FIXME type conversion
+        session.measureBandwidth((int) written);
     }
 
     public abstract int getAmountWritten();
